requirements.txt
openai
dotenv
python-dotenv
requests
pandas
aiohttp
async_timeout
tiktoken
faiss-cpu
tqdm
numpy
fastapi
uvicorn
